Fix integration tests

+++ _tests/integration/main_test.go
@@ -72,12 +72,13 @@ func Test_Integration(t *testing.T) {
 	cmd.Dir = rootPath
 	cmd.Env = append(os.Environ(), "GOPATH="+gopath)
 	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
 	err = cmd.Start()
 	if err != nil {
 		t.Fatalf("Unexpected error: %v", err)
 	}
 	defer func() {
-		if !cmd.ProcessState.Exited() {
+		if cmd.Process != nil && cmd.ProcessState != nil && !cmd.ProcessState.Exited() {
 			cmd.Process.Kill()
 		}
 	}()
